chore(react): updated formgroup playgorund, closes #12250
